config/property: Improve rendering of print
During startup, config properties are rendered, but sometimes in a
misleading way, especially durations:

```
redpanda.consumer_group_lag_collection_interval_sec:10000
```

The property type is seconds, with a value of 10. The value is
correct, but during rendering there is a conversion to milliseconds
and then the count is taken.

This commit improves the rendering, for example:

| before                                          | after                                         |
| ----------------------------------------------- | --------------------------------------------- |
| redpanda.aggregate_metrics:0                    | redpanda.aggregate_metrics:false              |
| redpanda.alive_timeout_ms:5000                  | redpanda.alive_timeout_ms:5000ms              |
| redpanda.group_offset_retention_sec:{604800000} | redpanda.group_offset_retention_sec:{604800s} |

Signed-off-by: Ben Pope <ben@redpanda.com>