archival: Bump 'paused_archivers' metric if ...
... the upload loop is blocked.
When the uploads are paused we're incrementing the 'paused_archivers'
metric to indicate that the archiver is not doing any job. The metric
is reset when the uploads are unblocked.

But there is another case when the uploads could be paused. The
'replica_validator' component is checking the state of the replica when
the 'ntp_archiver' is starting up. If it detects the anomaly it pauses
the uploads for the specific archver. This is needed to avoid burning
resources in situation when the data can't be accepted.

This commit changes this logic a bit so when the uploads are paused
due to anomaly the metric will indicate this.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit f4d3f3a9e1f09523d9402c02e30dfc1380ae3dd0)