net/batched_os: use gate instead of a plain boolean
Previously the `net::batched_output_stream` was using a plain boolean to
indicate that it was closed. The race condition might lead to a
situation in which a semaphore was accessed after the stream was
destroyed. This commit uses a gate to wrap the write action and wait for
it to finish before destroying the stream.

Fixes: CORE-9549

Signed-off-by: Michał Maślanka <michal@redpanda.com>