archival: Check replica validator with 'allow-gaps'
Add unit-test that creates a partition and simulates the situation when
the gap is created because of the upload pause with
'redpanda.remote.allowgaps' set to 'true'. The test then starts
'ntp_archiver' and checks that the upload loop is running and making new
uploads. The 'ntp_archiver' is allowed to upload because the gaps are
allowed.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit fe9b507f28b42711d6a43f6e42d90dd622e27ab2)