archival: Enable archiver manager by default
The archiver_manager component is currently disabled by default. When
it's disabled the `ntp_archiver` instances are managed by the
`cluster::partition`. When the partition is created it creates the
archiver and starts it. The archiver is never stopped until the
partition is stopped or some operations are performed. The operations
are: unsafe manifest reset and topic config change (e.g. compaction is
enabled).

During the leadership transfer the archiver is not stopped. Instead of
that 'prepare_transfer_leadership' and 'complete_transfer_leadership'
methods of the archiver are invoked. The methods are uploading the
recent version of the manifest and making sure that there are no orphan
segments in the cloud.

The 'archiver_manager' is a new component that manages 'ntp_archiver'
instances separately from the 'cluster::partition'. With
'archiver_manager' some features (e.g. force flusing archiver) will not
work at the moment. Also, the manager currently stops the archiver when
the leadership is lost instead of transferring the leadership.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>