various STM factories: use appropriate topic config revision
When partition state machines are created topic table may be already way
ahead and contain topic config for a newer revision of the topic, or
just lack it. This kind of race condition may happen if a topic is
deleted shortly after its partition is created. This change is to make
sure we use topic config from the same revision as the topic itself.