archival: Add 'get_rw_fence' method
Currently, the read-write fence is created by the 'ntp_archiver'
for every operation. The process is manual. The fence is consturcted
using the in-memory state of the STM and feature table.

This commit improves this by moving this logic into a dedicated method
called 'get_rw_fence'. The method constructs the fence and checks if the
'_mutex' is acquired. We don't want to run more than one concurrent
fibers. When this happens both/all fibers may acquire the same fence
value and only one of the fibers will be able to commit the changes
happily.

The 'ntp_archiver' is single threaded. No concurrency is allowed there
(with the exception of data movement). So the invariant is that only one
fence can be generated at a time. The '_mutex' check ensures that this
is the case.

Some unit-tests required change.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>