archival: Emit fence in reset-scrubbing-metadata
The method is invoked via the admin api. If it's invoked in parallel
with the upload the upload fails because there is an additional command
applied to the STM state. This makes the fence of the in-flight upload
invalid.

The commit adds a fence to the replicated scribber command and also
acquires units from the '_mutex' to avoid concurrency.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>