microbench: run rpbench_test in wrapper too
Use the same wrapper for run and test, as it increases coverage.

Previously we only use a wrapper for test, but after we introduced a
(more complicated) wrapper for run too, it makes sense to use it for
test as well, removing the small per-test wrapper we were creating
before.