r/consensus: do not use zero voter priority in maintenance mode
The maintenance mode should never completely prevent the leader of the
raft group to be elected. Even if the only possible leader candidate is
in maintenance mode. Change the leadership blocking logic to use the min
voter priority. This way the replica with blocked leadership can still
become a leader if it is the only eligible candidate.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit cb0e2f572cfbccf3e642faea06cbd67895d9839b)