cluster.py: always do crash checks
The "crash check" is where we scan the logs for Redpanda crashes at the
end of a test. Currently we only do this if a test fails. Some crashes
are caught by the bad log lines check, which occurs regardless of
pass or fail, but we should do the specific crash check in both cases
and before the bad log lines check, since it will produce a NodeCrash
error, which is more serious, handled differently in triage, etc.