cloud_storage: Add new spillover test
Add new unit test that reproduces the following situtation.
The manifest has a gap in offsets. The spillover command is added
alongside the truncate archive command (the same way its done by the
ntp_archiver during the first spillover operation for the partition).
The command sets archive start to be non-zero. In the normal case the
spillover command will succeed and the start offset will move forward
so the second command that moves the start offset will be applied
safely.

In this case though the spillover command fails because of the gap but
the archve start offset is moved forward. This creates an inconcistency
that triggers an assertion.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit 80636239f1576f2f8deaca97ba2a0b28445fdc16)