archival: Throw exception if spillover command..
..can't be applied. The exception is propagated to the caller
(ntp_archiver). The remaining commands in the record batch are not
applied.

The problem is that previously, when the STM couldn't apply the
spillover command it just skipped the command and applied remaining
commands in the record batch. Because of that it was possible to
introduce an inconsistency to the manifest.

Previous commit added a safety check that prevented the
'truncate_archive_init' command to be applied if it overshoots the start
offset. The situation that needs this occurs when the 'spillover'
command fails and then the 'truncate_archive_init' command is applied.
This commit pushes this further by forthing the STM to not apply any
commands after failed 'spillover' command.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit 549fdba6c4770a3c078af3d0119cb27c701effd0)