Fix e2e test_consumer_interruption flakiness
Decrease max.metadata.age to 1000 ms for this test,
so even if the consumer has its assignment revoked it will quickly
notice the topic being recreated and join again. With this fix, the test
passes locally 100% for me over 60 runs (vs about 50% fail before).

Details:

We delete and recreate a topic while a consumer is connected and look
for an error message which complains that the offset is wrong: this
occurs when the consumer starts consuming from the recreated topic after
already consuming from the original: in this case it try to consume from
a non-existent offset and get one of the expected errors.

The flakiness occurs because sometimes when the topic is deleted, the
consumer group handling will actually notice the topic deletion and
revoke the assignment of the topic to the consumer. In this case, the
consumer will need to wait max.metadata.age (5 minutes) in order to
“discover” the topic has been recreated. The result is that the consumer
doesn’t consume anything during the remaining period (about 15 seconds,
needed for the producer to produce 15,000 more messages at throughput
1000/s), so the expected errors (which originate either from the
on_consume callback or other message-consumed triggered code) do not
occur and instead we get a different error about the consumer losing
produced messages (which is also expected).

Which scenario occurs is really just a race between CG heartbeats vs
offset commit vs fetch requests: depending on which one “sees” the
deleted topic first, a different scenario develops.

To fix it, we increase the metadata query frequency, so it even if it
briefly loses the assignment, it will quickly notice the topic being
recreated and get assigned again.

Fixes CORE-9229.