storage: fix index state truncate overflow
CONFLICT:
- similar to https://github.com/redpanda-data/redpanda/pull/26024, the
  appropriate code in this branch is in segment_index instead of
  index_state
- because of this, the index_state unit test is removed

Similar to https://github.com/redpanda-data/redpanda/pull/25979, when an
index contains state that leads it to overflow, it is unreliable. In
this case, this could affect truncation, and result in us removing
incorrect index entries.

So, this makes the truncate avoid removing entries if this case is
detected, with the expectation that no entries are added that have
overflowed.

Note, this bug doesn't appear to be as severe as #25979 in that:
- suffix truncation is generally more rare
- the overflowing index problem is generally a problem with:
  - large segments, much larger than the default allowed size
  - compacted segments, which we expect to be relatively rare to
    suffix truncate with tiered storage

Nonetheless, some kind of fix seems appropriate because it's difficult
to reason about the overflow otherwise.

(cherry picked from commit e882d67ed669224fc7e713c423b3211eaf591eac)