c/a/upload: Add segment_collector::make_segment_upload_stream
Still uses make_upload_candidate under the hood to collect segments but feeds
the collected offset range into segment_upload::make_segment_upload.

Returns a segment_collector_stream as before. Intended as a mostly drop-in
replacement for make_upload_candidate_stream.

Note that this is not wired all the way back to the archiver yet. Implemented
here to test the interface and keep incremental diffs relatively small and
low risk.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>