cluster/types: make cluster::feature_update_cmd_data formattable
It's in the header, so it was probably intended to be implemented.

With it implemented we can print any command key which is helpful for
adding a debug print of all commands going through controller with their
types and keys. We're probably miles away from being able to print all
values.

(cherry picked from commit 1a00fd6f26f51962a9c5a749917bf72ae2712316)