kafka/server: Metadata: Do not return new topics before voting
When a topic is created, it is added to metadata_cache prior to
having a leader. The leader_epoch is 0, which can conflict with
a previous topic of the same name.

After voting, the term_id is set correctly, resulting in a correct
leader_epoch, which allows a consumer to detect the change.

This commit detects that a topic is new by not returning partitions
that have never had a leader; if there are no partitions, the topic
is rejected as unknown.

Signed-off-by: Ben Pope <ben@redpanda.com>