s/log_reader: Introduce optional read lock deadline
Set through log_reader_config. Currently we take a range lock over the segments
for the configured offset range, with no timeout.

When using log_readers in the TS upload loop, we're much more likely to race
against things like prefix truncation or segment rolls, so it's better to have
the option to configure a timeout.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>