s/log: Add non-blocking lock check in get_file_offset
If it fails, throw a timeout error so we skip across this upload attempt
and retry.

The main client of this API is offset_range_size, which takes an offset range
and aims to determine its size on disk. To this end, it first examines the
log to find the range of segments which contains the offset range boundaries
in its first and last element, respectively. Once the segment range is
determined, it is critical that the target segments do not change on disk for
the duration of the computation. To this end, offset_range_size first acquires
read locks for all the segments in its collected range.

Between the time we acquire this range of locks and the time we make a
subsequent call to get_file_offset as part of the range_size computation, it
is possible that a write_lock request for some segment has queued itself
behind the outstanding read lock. If we then try to acquire a read lock for
this segment in the file offset computation, we have a cyclic dependency
that leads to deadlock.

We must implement one of two courses of action here:
1. add a timeout to the lock acquisition in get_file_offset
2. check for this condition explicitly w/ a non-blocking lock call

It is unlikely that we ever approach the read lock limit for any such segment,
so the common case should be that any amount of waiting (as in (1)) is wasted
time. Instead, we can immediately fail the request w/ a 'synthetic' timeout
error, allow whatever write activity to take place as normal, and leave it to
the client of offset_range_size to try again.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>