ct: Use pimpl in cloud_topics/app
The 'app' now accepts a pointer to implementation (defined in the api.h)
in the c-tor. This is needed because the 'app' is created in the
'redpanda/application' and is a sharded service. And the sharded service
needs a concrete type, not the interface. So we can't create
'sharded_service<cloud_topics::api>' directly.

We also want to pass around the interface to avoid cyclic dependencies.
The 'app' implementation depends on 'cluster' because of the
'reconciler'. So passing 'app' to the cluster will create a cycle.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>