kafka: Change signature of the make_parition_proxy
The function now accepts 'cloud_topics::api' pointer or a sharded
service. The implementation of the function creates either
'replicated_partition' or 'cloud_topic_partition' based on topic
configuration. If the partition belongs to cloud topic the second
parameter is used to initialize 'cloud_topic_partition' instance.
Otherwise it's unused.

All callsites are updated. Note that in some cases the dataplane is not
propagated to the subsystem yet (e.g. datalake, WASM) so these
subsystems will not work with cloud topics.

Both fetch and produce paths are updated and are invoking
'make_partition_proxy' using cloud topics api. This enables cloud topics
read and write paths so end to end operation is possible. Note that
delete records request handler is not using cloud topics api yet. This
will be fixed in future commits but for now delete records request will
not work with cloud topics.

There is no impact on normal topics/partitions.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>