ct: Update read_request
The read_request was accepting log_reader_config as an input. The
underlying pipeline was using it to contruct a log reader that fetches
placeholders from the underlying partition.

This commit changes the read_request so it accepts a set of placeholder
batches as an input. The idea is that at the higher level the
coordinator will fetch placeholders from the control plane (underlying
partition) and use them to construct the read request. In this scenario
the read_pipeline is merely a data layer. The input is a set of
coordinates that the data layer uses to fetch data.

Note that in this model the timequery is a metadata lookup which is
performed in the control plane. The result of the operation is a set of
placehodlers from which the read_request can be constructed.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>