tests: use larger transaction timeout in describe producers test
The test was failing as the transactions that were expected to be
present when describing producers were already expired. Made the
transaction timeout larger to make sure the transactions are still
present when the producer list is being validated.

Fixes: CORE-10125

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 2548dd9866fd07892a80f0305f743976de7d1227)