log_eviction_stm: added monitoring eviction loop state
Log eviction is critical component of Redpanda as it ensures that
the log is kept within the configured retention limits. This state
machine is responsible for monitoring the log eviction events and
writing snapshots to ensure that the log is truncated correctly.
If the state machine loop is stalled or exits unexpectedly,
it can lead to log eviction not happening, which can cause the
log to grow indefinitely and eventually run out of disk space.
Therefore, we monitor the state of the loop and log an error if its state
is not updated for long time.

Signed-off-by: Michał Maślanka <michal@redpanda.com>