rptest: always run dbx test on aws and mark it as skipped on other clouds
This is a better default and much less likely for us to get a false
positive success run compared to previous implementation.

There is a chance that we'll want to skip dbx tests in some scenarios
even if we run with CLOUD_PROVIDER=aws. We'll address those as they
come.