ssx: Add execution_monitor class
The monitor is checking a set of site periodically and can detects
stalls and unexpected shutdown of the process.

The detector is using a periodic timer to check a set of pre-existing
callsties. The callsite is a simple data structure that contains a time
of the last checkpoint. There is a finite set of callsites.

The process invokes the 'checkpoint' method of the callsite while it
progresses.

There are also shutdown callsites. These are callsites which are reached
when the process is being stopped.

The execution monitor checks normal callsites and shutdown callsites
every stall-check-interval.

If all normal callsites weren't checkpointed within the stall interval
the stall is detected. If any shutdown callsite was checkpointed before
the 'expect_shutdown' was called the unexpected shutdown is detected.

In both cases the monitor invokes the callback provided by the user.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>