r/tests/raft_fixture: fix raft_node_instance::random_batch_base_offset
Previously it would misbehave if max is in a configuration or ghost
batch.
1) It expanded its search up to dirty_offset and returned a batch lying
entirely above max.
2) If it nevertheless failed to find a (non-ghost non-configuration)
batch it returned model::offset{} even though there could be some valid
batch before where it started to search.

The workaround is to expand search towards smaller offsets if we reached
the high end of the range.