c/a/uploader: Allow unstable reads in async_data_uploader
Skipping the raft layer allows us to bypass visible offset checks when
constructing a log_reader.

This is useful in situations where the desired upload bounds are a priori known
to be safe but may not precisely match the consumer visible offset. e.g. when
doing time-based uploads, direct segment collection, or in test.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>