`storage`: add `segment_reader::fsize()`
Currently, `segment_reader::stat()` is only called in places where
we require the file size of the underlying `segment` file.

To offer a way of getting the size of a file on disk without needing to
issue a syscall which returns all of the `stat(2)` data, add `fsize()`.

(cherry picked from commit 2323dad98642d951c217405068339dfc79b39dd7)