`storage`: add `cmp_idx_size_after` to `compaction_result`
As a way to pass the new size of the rewritten `.compaction_index` file
between functions in the compaction util call stack.

This will be used in a future commit to manually set the cached size
of the `.compaction_index` within a `segment` in memory.

(cherry picked from commit ce406e20354da12676b4d606317f42a663bc42b6)