`storage`: use appender sizes if possible in `segment::persistent_size()`
Before a `segment` is closed and `appender`s are still owned, they can be
used to provide an estimate of the underlying file size on disk without
needing to make a call to `ss::file_size()`.

Use these in `persistent_size()` to cut down on the number of syscalls
to `ss::file_size()` while a `segment` is still open (and constantly
invalidating its cached sizes through various `segment` operations).