config: added property to allow enabling batch cache for consumer offsets
It may sometimes be beneficial for troubleshooting to enable
`__consumer_offsets` topic batch cache. The cache is disabled for this
topic as it is read only once when applying batches to the stm. The
batches written to the topic are usually very small. Skipping storing
them in the cache relieve the cache index pressure.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit d402ebdddf2e98795a706533a568f8b2db442d41)