`storage`: add `segment_appender::size_bytes()`
This function seeks to report the actual size of the underlying file on
disk. `file_byte_offset()` for an open `segment_appender` will report
the current offset for the cursor, but since we `fallocate()` the file
up front, the actual file size is represented by `_fallocation_offset`.

Of course, the size reported by `file_byte_offset()` and `size_bytes()` will
be identical for a closed `segment_appender`. But for querying the size of
a segment with an open `appender`, `size_bytes()` should be used instead.