dt/ts: Add rate limiting in ts_model_test
In the new log_reader-based upload collection path, remote segment sizes can
and do approach the overall remote segment size target on initial upload if
the throughput of the source partiton allows it. By dialing down the produce
rate, we wind up with smaller remote segments, allowing adjacent segment
merging to occur in model tests that look for it.