cs/remote_partition: do not try to enter gate in noexcept constructor
`remote_partition` reader constructor is marked as `noexcept`.
Previously the gate holder was created in the constructor. When a holder
is constructed from the gate it may throw `gate_closed_exception` when
the gate is closed. This lead to an assertion being trigger as the
constructor is not allowed to throw exceptions.

Fixes: CORE-12454

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit cd411306de44675828a42f3ae9ea01981a7e9da2)