cloud_io: Detect shutdow
The http client returns timeout error if the connection attempt is
aborted. This creates an issue when the client is aborted during the
shutdown. In this case the shutdown error is interpreted as a retriable
error by the cloud storage client. The cloud_io::remote is retrying the
operation in this case.

In order to avoid this this commit adds a check which examines the state
of the abort source and if the abort was requested it converts the error
into non-retriable type.

NOTE: it is possible to solve this problem by adding a new 'shutdown'
error code to the http client. This is tempting but also dangerous. The
clients are interrupted not only during the shutdown but during the
normal operation (e.g. in case of 500 error). Some code may interpret
this error as a true shutdown and cause disruption. For instance, the
background loop of some subsystem may exit early after catching this
error.

To avoid this problem the error is handled in the cloud_io::remote
without introducing the special type for shutdown.

Signed-off-by: Evgeny Lazin <4lazin@gmail.com>
(cherry picked from commit cfd68d6ae317cd04c11f9cf9b0b2306b9e66b102)