dt/remote_label_test: Set a sensible cs_segment_size_target
If unset, we default to log_segment_size, which is tuned way down to 1KiB in
this test, leading to lots of very small uploads.

On rehydration of the cloud cache after swapping objects in the
concurrent_consume test, segment download was consistently hitting BlobNotFound
errors from the object store.

TODO: Needs some investigation to pin down. Though, notably, in the old world
the upload sizes were more reasonable.