c/self_check: fix list pagination handling
Azure ABS treats pagination quite flexibly and the ABS API sometimes
responds to List Blob requests with an empty list of results + a
continuation token for more results, prompting the user to try again for
more data.

This lead to a bug in redpanda's self check logic, where the self check
would fail when as part of one of the checks, it saw that the incomplete
List Blob response did not contain the earlier uploaded object that it
would have expected to see in the List Blob response.

This was highlighted by flakiness in Azure CDT.

Ref: https://learn.microsoft.com/en-us/answers/questions/1330367/rest-api-not-returning-blobs-within-container
(cherry picked from commit 2a60f9f182ffea4d3e3265ead2995713c2b98f9a)