sec/authorizer: support multi-resource checks
The GET /schemas/ids/{id} endpoint is authorized based on a set of
resources instead of a single resource.

To support this, this commit introduces an `any_authorized` method on
the authorizer object which allows access if the user has access to any
of the resources, or if the user is a superuser.

The `auth_result` object returned by `any_authorized` may not contain a
resource_name if none were given in the argument. In this case, audit
logging uses a hardcoded resource name to highlight that the action was
authorized/denied based on an empty set of resources.

The implementation ensures that the `authorization_result` metric is
only incremented once for a single call of `any_authorized`, since this
is a single authroization check and we want to avoid including in the
probe the results of intermediary authorization checks.