k/s/group: address inconsistency in group subscription
Compared to the reference implementation, the redpanda
logic on whether a group is subscribed is too restrictive.
Specifically in the fallback path, if the _subscriptions member
is not set. We consider all such groups as subscribed.

Updated fallback logic to consider the group subscribed if the
protocol type is consumer group. Otherwise, it falls back to
considering the group unsubscribed. This logic matches the
reference implementation.

For the full discussion see CORE-12338

(cherry picked from commit 8a32acbefb61262da261cc5b882134c5e7559196)