tests: exclude the JSON fuzz test from CI / ...
This tags the json fuzz test as manual, so it won't run in CI or get
matched by wildcards like `...`. A fuzz test is not meant to be run in
CI like a unit test.

Note that if we develop a good corpus then the fuzz test can be run over
the corpus as a regression test. libfuzzer has a mode for this.