`storage`: remove `ntp` from `is_compactible_control_batch()` path
This was only in place to ensure `tx_fence` batches were considered
compactible in the `__consumer_offsets` topic and not in user topics.

However, `tx_fence` batches should be considered safely removable during
compaction, as the presence of the transactional bit in a Raft data batch
can accomplish the same "fencing" marker as the `tx_fence` batch itself.