`storage`: rename `is_compactible_control_batch()` to `is_removable...`
"Compactible", as a concept, should be reserved for the idea that a record
is discardable due to the existence of a later record with the same key
and a higher offset in the `log`.

Since we are discarding these batches unconditionally without any
consideration or requirement for the above condition, `is_removable()`
is a better term to distinguish the nuance in behavior here.