cluster/node_status_backend: transport memory leak
If a node is suspended, connections to it will remain alive despite
never making progress. If rpc's continue to be written to this
connection, an unbounded amount of memory builds in the transport buffer
in seastar.

The solution here is the same as that of heartbeat_manager. If a
sufficient number of subsequent timeouts is observed, reset the
connection.

(cherry picked from commit 55d37b8ad33564e0280cc1a72ec285bf777cfc79)