raft/buffered_protocol: fix race between GC and buffering appends
It is possible that the incoming append is buffered in an append queue
that just got GC-ed due to scheduling points. This was manifesting as
dangling future of append_entries RPC and its down stream effects in
consensus like timed out replication and forever held raft gate
blocking shutdown.

This commit ensures that an inflight, to be buffered append only lands
in a queue that is not a candidate for GC.

(cherry picked from commit c2e803fe3debf0ca5ae7d42b84a972305b55aab6)