dt/archival: Decrease manifest upload interval to avoid race
This applies to 'ArchivalTest.test_all_partitions_leadership_transfer', where
validation checks that the end offset of the topic matches the end of the
partition manifest in the bucket.

In some cases, the test NTP stabilizes just after a partition manifest upload,
which means that we will have to wait at least a minute before the manifest is
in sync. This in turn puts us over the validation timeout for the test.

So this commit cuts the manifest_max_upload_interval in half for this test
case only. The alternative is to increase the test timeout, but that would
increase the runtime of the test in these outlier cases.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>
(cherry picked from commit 010e6d8bb659bc88e40ead97c28587fc5680f438)