controller: Skip FPM truncation if TS is not enabled
Fast partition movement works by setting the learner's raft start offset to
a safe prefix truncation of the local log, based on initial retention configs
and some calculations to determine an offset below which all data are recoverable
from cloud storage. Raft data above this offset are transferred directly,
whereas data below this offset (assumed to be recoverable from cloud) are not.

Prior to this change, this computation could produce an unsafe initial start
offset for the learner if archival_metadata_stm::max_removable_local_log_offset
exceeds the last uploaded offset, which can occur if archival is disabled.

Note that this difference between max_removable and last uploaded is by design
since automatic truncation is usually predicated on cleanup policy, and archival
being switched off should not block cleanup in the common case.

FPM is a special case in the sense that the "truncation" is not for cleanup
as such, but rather an omission of backed-up data from an expensive bulk
transfer. Which is to say it can and does occur even if the source topic is
compact-only.

The net result of this is that archival_metadata_stm reports that it is safe
to remove up to offset::max, and, as a result, FPM truncates the raft transfer
up to whatever removable offset is reported by the other stms. Any data produced
between the last cloud storage upload and this offset are lost.

This commit introduces checks to

- prevent FPM from adjusting the learner start offset above the start of the
  local log for any partition w/ TS disabled or w/ uploads paused
- prevent FPM from operating on internal topics
- ensure that FPM  won't truncate past the safe offset in general as reported
  by archival metadata stm. This is mostly a safeguard, as in most cases
  max_removable_local_log_offset should return the correct thing.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>
(cherry picked from commit 3dc8d349563a192d63d2cfb16df93ebc4d3d03de)

Conflicts:
- max_collectible_offset was renamed to max_removable_local_log_offset
- local naming in controller backend
- config::cloud_storage_enable_segment_uploads not yet implemented
- turn off upload pause ducktape test