cloud_storage: hold gate in remote
CONFLICT:
- cloud_io::remote constructor has one fewer arg, had to fix remote_test

Previously we held a gate in cloud_io::remote, but not
cloud_storage::remote. This led to a heap-use-after-free where
cloud_storage::remote is successfully destructed, but cloud_io::remote
is in the middle of a retry -- in this case, it calls an input callback
that relies on cloud_storage::remote::_probe.

This commit fixes this by adding a gate and abort source to
cloud_storage::remote and holding the gate in all async calls.

(cherry picked from commit d1eb7f781463d963bacc2b970fa457c41bf30b16)