raft/voter_priority_tracker: fix first leader nepo
On creation of a group, the first node_id in the replica set is supposed
to preferentially be selected as leader. This PR fixes a bug where on
creation, the initial 'target_priority' was set to zero instead of MAX,
which allowed for all nodes to vote for themselves immediately, instead
of waiting a few rounds for the target priority to fall.

(cherry picked from commit 53d94e89507d6641a6fc2badc63999e9fccd571d)