`cloud_storage`: fall back to main manifest search in `get_term_last_offset()`
When we fail to find an entry in the spillover map via
`get_spillover_upper_bound_by_term()`, we need to fall back to searching
the main manifest for the next `segment` term to properly respond to
a `OffsetForLeaderEpoch` query.

Consider the following edge case:

Populate 2 spillover manifests in the `partition_manifest`'s `spillover_map`.
The desired term 1, is last entry in the spillover manifests, i.e there are
no higher term entries in `get_segment_term_column`, and the main manifest
contains no segments with a lower term than the desired term. This can be
illustrated as:

```
Main manifest: [2, 3]
Spillover map: [[0], [1]]
```

`get_spillover_upper_bound_by_term()` will return `std::nullopt`, since no
higher bound for the desired term 1 was found in the 2 entries of the
spillover map. Previously, this case would erroneously be treated as an
early return of `std::nullopt`. However, we still need to consider the main
manifest in this case, as the base offset of the `segment` with term 2
present in the main manifest will give us our proper response to the query.

(cherry picked from commit e3d78c01c6283dc2533500807f9c91e1eced48fa)