archival: abort mutex wait if adjacent segment merger is being stopped
This should fix archival stop hang when an upload is taking longer time.
With this commit, the mutex wait is aborted immediately and we can
proceed stopping the archiver itself which will abort the upload
altogether.

Before this commit, we have to wait for tiered storage upload to
finish. Or for a housekeeping operation to finish.

(cherry picked from commit 9b92a5056b7bc1c2c9c5a51aa44207316b43953c)