bazel: support flaky attribute on tests
Bazel has a flaky attribute which changes some behavior when set on
a test (e.g., the test may be automatically retried), but we do not
support it in our btest and gtest macros.

Add support for it and pipe it through to bazel.

Fixes CORE-12895.

(cherry picked from commit 84df06ebf5d92868e7fd380ed32486179ea291ef)