c/data-migrate/worker: rework management of works for the same NTP
Previously work tracking logic was flawed. When for an NTP a work was
already running and backend requested another work (e.g. for a different
stage or for a new migration), it was handled inappropriately.
1) Existing running work was only aborted by triggering abort source,
but not waited to actually stop. As a result, works on the same NTP
could run concurrently, in particular in the wrong order. This might
lead to incorrect results, e. g. topic writes unblocked then blocked
instead of the other way round.
2) Work info, which is supplementary data a work uses, was overwritten
by the new one. The old work which was still running might access
deallocated or reused memory where the old work info was, thus
triggering an UB.
3) Per-work abort sources were not in use, only main one was.

Reorganised logic:
1) Allow no more than one running work per NTP and no more than one
queued. Semaphores and mutexes would not be able to help with it as they
maintain a potentially long queue of waiters. We allow only one waiting.
2) Store its belongigns separately from the one requested if they are
different.
3) Use both main and individual abort sources.

(cherry picked from commit d9f07a7bc2498af20c64deef893585cd34c2f482)