ct/l1: expose domains in metastore interface
In order to ensure that data is not shared across domain boundaries,
writers that use the metastore (e.g. reconciler, compaction, leveler)
will need to be domain-aware.

This adds a "domain_id" as a part of the metastore public interface,
with a dummy numeric value being used in the simple_metastore. I expect
there to be a more complex grouping of topics to domains in the future.
I also expect that L1 topic partitions will each manage data for a
single domain, and that domain_id will be used to route requests to the
appropriate L1 topic partition.

For now though, it's just exposed so that code authors can do someting
like (roughly):

```
map<domain_id, object_builder> objects;
for (const auto& tp : topic_partitions) {
    auto domain_id = metastore.get_domain_id(tp);
    objects[domain_id].add(...data from tp...);
}
for (auto& [id, ob] : objects) {
    metastore.add_objects(id, ob.build());
}
```